:recycle: Refactor RenderGrid into multiple components
