--- src/components/RenderGrid/RenderGrid.css
@@ -1,24 +1,0 @@
-.canvas-container {
-    width: 100vw;
-    height: 100vh;
-}
-
-canvas {
-    width: 100vw !important;
-    height: 100vh !important;
-    display: block;
-}
-
-.rotate-left-button {
-    position: absolute !important;
-    z-index: 1000;
-    right: 0;
-    margin: 90px 16px 0 0 !important;
-}
-
-.rotate-right-button {
-    position: absolute !important;
-    z-index: 1000;
-    right: 0;
-    margin: 160px 16px 0 0 !important;
-}
